+++ dbms/src/Storages/KVStore/tests/gtest_raftstore_v2.cpp
@@ -149,7 +149,7 @@ CATCH
 TEST_F(RegionKVStoreV2Test, KVStoreExtraDataSnapshot1)
 try
 {
-    auto ctx = TiFlashTestEnv::getGlobalContext();
+    auto & ctx = TiFlashTestEnv::getGlobalContext();
     proxy_instance->cluster_ver = RaftstoreVer::V2;
     ASSERT_NE(proxy_helper->sst_reader_interfaces.fn_key, nullptr);
     UInt64 region_id = 1;
@@ -244,7 +244,7 @@ CATCH
 TEST_F(RegionKVStoreV2Test, KVStoreExtraDataSnapshot2)
 try
 {
-    auto ctx = TiFlashTestEnv::getGlobalContext();
+    auto & ctx = TiFlashTestEnv::getGlobalContext();
     proxy_instance->cluster_ver = RaftstoreVer::V2;
     UInt64 region_id = 1;
     TableID table_id;
@@ -354,7 +354,7 @@ CATCH
 TEST_F(RegionKVStoreV2Test, KVStoreExtraDataSnapshot)
 try
 {
-    auto ctx = TiFlashTestEnv::getGlobalContext();
+    auto & ctx = TiFlashTestEnv::getGlobalContext();
     proxy_instance->cluster_ver = RaftstoreVer::V2;
     UInt64 region_id = 1;
     TableID table_id;
@@ -465,7 +465,7 @@ CATCH
 TEST_F(RegionKVStoreV2Test, KVStoreSingleSnapCancel)
 try
 {
-    auto ctx = TiFlashTestEnv::getGlobalContext();
+    auto & ctx = TiFlashTestEnv::getGlobalContext();
     proxy_instance->cluster_ver = RaftstoreVer::V2;
     ASSERT_NE(proxy_helper->sst_reader_interfaces.fn_key, nullptr);
     ASSERT_NE(proxy_helper->fn_get_config_json, nullptr);
@@ -532,7 +532,7 @@ CATCH
 TEST_F(RegionKVStoreV2Test, KVStoreSingleSnap1)
 try
 {
-    auto ctx = TiFlashTestEnv::getGlobalContext();
+    auto & ctx = TiFlashTestEnv::getGlobalContext();
     proxy_instance->cluster_ver = RaftstoreVer::V2;
     ASSERT_NE(proxy_helper->sst_reader_interfaces.fn_key, nullptr);
     ASSERT_NE(proxy_helper->fn_get_config_json, nullptr);
@@ -616,7 +616,7 @@ CATCH
 TEST_F(RegionKVStoreV2Test, KVStoreSingleSnap2)
 try
 {
-    auto ctx = TiFlashTestEnv::getGlobalContext();
+    auto & ctx = TiFlashTestEnv::getGlobalContext();
     proxy_instance->cluster_ver = RaftstoreVer::V2;
     ASSERT_NE(proxy_helper->sst_reader_interfaces.fn_key, nullptr);
     ASSERT_NE(proxy_helper->fn_get_config_json, nullptr);
@@ -676,7 +676,7 @@ CATCH
 TEST_F(RegionKVStoreV2Test, KVStoreSingleSnap3)
 try
 {
-    auto ctx = TiFlashTestEnv::getGlobalContext();
+    auto & ctx = TiFlashTestEnv::getGlobalContext();
     proxy_instance->cluster_ver = RaftstoreVer::V2;
     ASSERT_NE(proxy_helper->sst_reader_interfaces.fn_key, nullptr);
     ASSERT_NE(proxy_helper->fn_get_config_json, nullptr);
@@ -736,7 +736,7 @@ CATCH
 TEST_F(RegionKVStoreV2Test, KVStoreSingleSnap4)
 try
 {
-    auto ctx = TiFlashTestEnv::getGlobalContext();
+    auto & ctx = TiFlashTestEnv::getGlobalContext();
     proxy_instance->cluster_ver = RaftstoreVer::V2;
     ASSERT_NE(proxy_helper->sst_reader_interfaces.fn_key, nullptr);
     ASSERT_NE(proxy_helper->fn_get_config_json, nullptr);
@@ -809,7 +809,7 @@ CATCH
 TEST_F(RegionKVStoreV2Test, KVStoreSingleSnap5)
 try
 {
-    auto ctx = TiFlashTestEnv::getGlobalContext();
+    auto & ctx = TiFlashTestEnv::getGlobalContext();
     proxy_instance->cluster_ver = RaftstoreVer::V2;
     ASSERT_NE(proxy_helper->sst_reader_interfaces.fn_key, nullptr);
     ASSERT_NE(proxy_helper->fn_get_config_json, nullptr);
@@ -869,7 +869,7 @@ CATCH
 TEST_F(RegionKVStoreV2Test, KVStoreSingleSnap6)
 try
 {
-    auto ctx = TiFlashTestEnv::getGlobalContext();
+    auto & ctx = TiFlashTestEnv::getGlobalContext();
     proxy_instance->cluster_ver = RaftstoreVer::V2;
     proxy_instance->proxy_config_string = R"({"raftstore":{"snap-handle-pool-size":3}})";
     KVStore & kvs = getKVS();
